building out some of the core foundational logic. more to come soon.
I'm currently building out the core foundational logic for the traverser worker, tasks, queue, and queue publisher structs. Just need to take a step back and work on this more later with a fresh mind so I can see more clearly what's missing in the gaps and fill them in. Once that comes, we can build out some basic endpoints and start testing!